requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
transformers
torch
httpx
Pillow
pytest
